src/react-native-explorer/requirements.txt
mcp>=1.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
pyyaml>=6.0
Pillow>=10.0.0
rich>=13.0.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
python-dotenv>=1.0.0
